+++ types/index.ts
@@ -29,7 +29,7 @@ export type { DuCheckboxVariant,DuCheckboxSize,DuCheckboxProps } from '../compon
 export type { DuFieldsetProps } from '../components/DataInput/du-fieldset/du-fieldset.types'
 export type { DuFileInputEmit,DuFileInputProps } from '../components/DataInput/du-file-input/du-file-input.types'
 export type { DuFilterValue,DuFilterEmit,DuFilterButtonArgs,DuFilterItem,DuFilterProps } from '../components/DataInput/du-filter/du-filter.types'
-export type { DuInputFieldVariant,DuInputFieldSize,DuInputFieldType,DuInputFieldValidatorProps,DuInputFieldProps } from '../components/DataInput/du-input-field/du-input-field.types'
+export type { DuInputFieldVariant,DuInputFieldSize,DuInputFieldType,DuInputFieldValidatorProps,DuInputFieldModelModifier,DuInputFieldProps } from '../components/DataInput/du-input-field/du-input-field.types'
 export type { DuLabelInputValidatorProps } from '../components/DataInput/du-label-input-validator/du-label-input-validator.types'
 export type { DuLabelProps } from '../components/DataInput/du-label/du-label.types'
 export type { DuRadioVariant,DuRadioSize,DuRadioProps } from '../components/DataInput/du-radio/du-radio.types'